+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # lignin-html Changelog
 
+## next
+
+TODO: Date
+
+* **Breaking:**
+  * Removed "remnants" features (always enabled now)
+  * Upgraded `lignin` dependency to 0.0.2
+
 ## 0.0.1
 
 2020-10-02